requirements.txt
fastapi
sqlalchemy
pydantic
sqlmodel
psycopg2-binary
uvicorn
PyJWT
bcrypt
passlib
boto3
python-multipart

requests
pytest
pytest-cov

black
isort
flake8